UI: Delete OBSQTDisplays before dialogs get closed
On X11, there can be a segfault that occurs when a dialog window gets
closed but there is still a OBSQTDisplay object attached to the
closed window. Currently, these OBSQTDisplay objects are deleted via
an inherint deleteLater() method, which opens time for the OBS
graphics thread to attempt to switch contexts to the invalid window,
which causes the segfault. By deleting the OBSQTDisplay before the
dialog window gets closed, we avoid this problem.